cli/requirements.txt
absl-py
google-api-python-client
google-cloud-aiplatform
kubernetes
